complete core refactory of the activitygen.py script.
